CONFETI-39 refactor: 응답 타입을 swagger 에서 자동으로 매핑하도록 수정 - docs 인터페이스를 docs 패키지 하위로 옮김
